+++ static/js/project.js
@@ -30,4 +30,16 @@ $(function() {
         $("#transfer-confirm").prop("disabled", !this.value);
     });
 
+    $("a[data-revoke-key]").click(function() {
+        $("#revoke-key-type").val(this.dataset.revokeKey);
+        $("#revoke-key-modal .name").text(this.dataset.name);
+        $("#revoke-key-modal").modal("show");
+    })
+
+    $("a[data-create-key]").click(function() {
+        $("#create-key-type").val(this.dataset.createKey);
+        $("#create-key-form").submit();
+    })
+
+
 });